+++ packages/cli/lib/index.js
@@ -1,5 +1,0 @@
-const testPages = require('./axe-test-urls');
-const utils = require('./utils');
-const saveOutcome = require('./save-outcome')
-
-module.exports = { saveOutcome, testPages, utils };