+++ packages/react/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.2.1](https://github.com/dequelabs/axe-core-npm/compare/v4.2.0...v4.2.1) (2021-05-19)
+
+### Bug Fixes
+
+- update axe-core to 4.2.1 ([#254](https://github.com/dequelabs/axe-core-npm/issues/254)) ([9d90185](https://github.com/dequelabs/axe-core-npm/commit/9d9018525a4d799f6d763d0329f05ccbfd20dbe4))
+
+# [4.2.0](https://github.com/dequelabs/axe-core-npm/compare/v4.1.1...v4.2.0) (2021-05-05)
+
+### Features
+
+- update axe-core to 4.2.0 ([#240](https://github.com/dequelabs/axe-core-npm/issues/240)) ([4e8f7fe](https://github.com/dequelabs/axe-core-npm/commit/4e8f7fee9db09fb56f91ea34f9984be66a29033e))
+
+## [4.1.1](https://github.com/dequelabs/axe-core-npm/compare/v4.0.0...v4.1.1) (2021-01-28)
+
+### Bug Fixes
+
+- **react:** Add example using Next.js ([#109](https://github.com/dequelabs/axe-core-npm/issues/109)) ([bd009d9](https://github.com/dequelabs/axe-core-npm/commit/bd009d9578d9b2704e258ade932fdbbe0a7a571a)), closes [#103](https://github.com/dequelabs/axe-core-npm/issues/103)
+- **react:** handle undefined config ([#183](https://github.com/dequelabs/axe-core-npm/issues/183)) ([6326e13](https://github.com/dequelabs/axe-core-npm/commit/6326e131bf62968d5a8180595c1b0b049844ca05))
+- **react:** Prevent config TypeError ([708c463](https://github.com/dequelabs/axe-core-npm/commit/708c46366d5f2069a4f25238beb32299abc962b9))
+
+### Features
+
+- **react:** add support for runOnly. ([#101](https://github.com/dequelabs/axe-core-npm/issues/101)) ([cfadde3](https://github.com/dequelabs/axe-core-npm/commit/cfadde3eebf74d26e0fb3768b84e953e92f0bf99))
+
 # [4.1.0](https://github.com/dequelabs/axe-core-npm/compare/v4.0.0...v4.1.0) (2020-11-20)
 
 ### Bug Fixes